bl/requirements.txt
# asgiref==3.10.0
# blinker==1.9.0
# certifi==2025.11.12
# cffi==2.0.0
# channels==4.3.1
# charset-normalizer==3.4.4
# click==8.2.1
# colorama==0.4.6
# cryptography==46.0.3
# defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-silk==5.4.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
django_celery_results==2.6.0
django-redis==6.0.0
# djoser==2.3.3
# flake8==7.3.0
# Flask==3.1.1
# gprof2dot==2025.4.14
# idna==3.11
# itsdangerous==2.2.0
# Jinja2==3.1.6
# MarkupSafe==3.0.2
# mccabe==0.7.0
# oauthlib==3.3.1
pillow==11.3.0
# pycodestyle==2.14.0
# pycparser==2.23
# pyflakes==3.4.0
# pyjokes==0.8.3
# PyJWT==2.9.0
# python3-openid==3.2.0
# requests==2.32.5
# requests-oauthlib==2.0.0
# social-auth-app-django==5.6.0
# social-auth-core==4.7.0
# sqlparse==0.5.3
# tzdata==2025.2
# urllib3==2.6.2
# Werkzeug==3.1.3
